// api/bindings/Activity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Room } from "./Room";
import type { StaffMember } from "./StaffMember";

export interface Activity { id: string, courseCode: string, week: number, start: string, end: string, title: string, summary: string, staffMembers: Array<StaffMember>, studentGroups: Array<string>, rooms: Array<Room>, }